+++ packages/components/molecules/f-mega-modal/jest.config.js
@@ -37,9 +37,5 @@ module.exports = {
         './test/specs/component'
     ],
 
-    testURL: 'http://localhost/',
-
-    setupFilesAfterEnv: [
-        '../../../../jest.setup.js'
-    ]
+    testURL: 'http://localhost/'
 };