+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout-selectors.js
@@ -11,6 +11,10 @@ export const GUEST_CHECKOUT_HEADER = '[data-test-id="guest-checkoutHeader"]';
 export const PRE_ORDER_WARNING = '[data-test-id="warning-pre-order"]';
 export const CHECKOUT_ERROR_MESSAGE = '[data-test-id="checkout-issue-modal"]';
 export const RETRY_BUTTON = '[data-test-id="redirect-to-menu-button"]';
+export const ERROR_PAGE_COMPONENT = '[data-test-id="checkout-error-page-component"]';
+export const ERROR_PAGE_HEADING = '[data-test-id="checkout-error-page-heading"]';
+export const ERROR_PAGE_DESCRIPTION = '[data-test-id="checkout-error-page-description"]';
+export const ERROR_PAGE_IMAGE = '[data-test-id="checkout-error-page-image"]';
 
 export const FIELDS = {
     firstName: {
@@ -50,14 +54,4 @@ export const FIELDS = {
         input: '[data-test-id="user-note"] textarea',
         error: ''
     }
-};
-
-export default {
-    CHECKOUT_COMPONENT,
-    ORDER_TIME_DROPDOWN,
-    ORDER_TIME_DROPDOWN_OPTIONS,
-    USER_NOTE_INPUT,
-    GO_TO_PAYMENT_BUTTON,
-    SWITCH_USER_LINK,
-    FIELDS
 };