+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-delivery.component.spec.js
@@ -5,7 +5,7 @@ const Checkout = require('../../../test-utils/component-objects/f-checkout.compo
 const checkout = new Checkout();
 
 describe('f-checkout "delivery" component tests', () => {
-    before(() => {
+    beforeEach(() => {
         const checkoutData = {
             type: 'delivery',
             isAuthenticated: true,