+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout.component.spec.js
@@ -3,7 +3,7 @@ const Checkout = require('../../../test-utils/component-objects/f-checkout.compo
 const checkout = new Checkout();
 
 describe('f-checkout component tests', () => {
-    before(() => {
+    beforeEach(() => {
         const checkoutData = {
             type: 'delivery',
             isAuthenticated: true,