+++ packages/services/f-braze-adapter/jest.config.js
@@ -1,6 +1,5 @@
 module.exports = {
     setupFilesAfterEnv: [
-        'jest-extended',
-        '../../../jest.setup.js'
+        'jest-extended'
     ]
 };